Paper2D: Add a visual preview of the current ink on the tile map painting brush using a preview component
[CL 2404317 by Michael Noland in Main branch]

+++ Engine/Plugins/2D/Paper2D/Source/Paper2D/Classes/PaperTileMap.h
@@ -109,7 +109,7 @@ public:
 	// End of UObject interface
 
 
-	FVector GetTilePositionInLocalSpace(int32 TileX, int32 TileY, int32 LayerIndex = 0) const;
+	FVector GetTilePositionInLocalSpace(float TileX, float TileY, int32 LayerIndex = 0) const;
 
 	FBoxSphereBounds GetRenderBounds() const;
